pyproject.toml
[project]
name = "ml-data-project"
version = "0.1.0"
description = "A Python project for data analysis and machine learning tasks"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "plotly>=5.0.0",
    "scipy>=1.7.0",
    "sqlalchemy>=1.4.0",
    "pymysql>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "sqlfluff>=2.0.0",
    "pre-commit>=2.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"